[#265] test: fix tests
